ai_lead_agent/requirements.txt
feedparser
beautifulsoup4
langchain
langchain-groq
rapidfuzz
ddgs
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
openpyxl
apscheduler
pandas
numpy
python-dotenv
urllib3
requests
lxml
anthropic